Added more API calls, Added Weather, Profile, Location, and Accounts classes
Left off working on the Profile Class, I think I want to restructure this.

We should have a single Client class. This client class manages everything about that client, including their profiles, etc. Then we simply have an array of clients in a Server class that has helper functions for accessing and managing a client.